requirements.txt
numpy==2.3.2
numpy-stl==3.2.0
pillow==11.3.0
